Remove PatriciaTriePath and tests
